feat: break homepage  in components
