+++ .vscode/settings.json
@@ -1,0 +1,20 @@
+{
+  "editor.tabSize": 2,
+  "editor.detectIndentation": false,
+  "js/ts.preferences.autoImportSpecifierExcludeRegexes": [
+    // useRouter should be imported from `next/navigation` instead of `next/router`
+    "next/router",
+    // give priority for Link to next/link instead of lucide-react
+    "lucide-react",
+    // Sentry is imported with `import *`
+    "@sentry/nextjs"
+  ],
+  "search.exclude": {
+    "pnpm-lock.yaml": true
+  },
+  "editor.codeActionsOnSave": {
+    "source.fixAll.ts": "explicit",
+    "source.addMissingImports": "explicit",
+    "source.fixAll.eslint": "explicit"
+  }
+}
