+++ src/assets.ts
@@ -1,0 +1,4 @@
+export const Assets = {
+    aiFace: "/images/ai-face.jpg",
+    userFace: "/images/user-face.jpg"
+} as const
